# Simplify `get_buffer_nbytes` `min_size` check