--- a/tests/test_libs_utils.py
+++ b/tests/test_libs_utils.py
@@ -40,18 +40,16 @@ def test_get_buffer_data_builtins(buffer):
 @pytest.mark.parametrize("buffer", builtin_buffers)
 def test_get_buffer_nbytes_builtins(buffer):
     nbytes = memoryview(buffer).nbytes
-    result = get_buffer_nbytes(buffer, check_min_size=None, cuda_support=True)
+    result = get_buffer_nbytes(buffer, cuda_support=True)
     assert result == nbytes
 
     with pytest.raises(ValueError):
-        get_buffer_nbytes(
-            memoryview(buffer)[::2], check_min_size=None, cuda_support=True
-        )
+        get_buffer_nbytes(memoryview(buffer)[::2], cuda_support=True)
 
     # Test exceptional cases with `check_min_size`
-    get_buffer_nbytes(buffer, check_min_size=nbytes, cuda_support=True)
+    get_buffer_nbytes(buffer, min_size=nbytes, cuda_support=True)
     with pytest.raises(ValueError):
-        get_buffer_nbytes(buffer, check_min_size=(nbytes + 1), cuda_support=True)
+        get_buffer_nbytes(buffer, min_size=(nbytes + 1), cuda_support=True)
 
 
 array_params = [
@@ -94,8 +92,8 @@ def test_get_buffer_nbytes_array(xp, shape, dtype, strides):
     xp, arr, iface = create_array(xp, shape, dtype, strides)
 
     if arr.flags.c_contiguous:
-        nbytes = get_buffer_nbytes(arr, check_min_size=None, cuda_support=True)
+        nbytes = get_buffer_nbytes(arr, cuda_support=True)
         assert nbytes == arr.nbytes
     else:
         with pytest.raises(ValueError):
-            get_buffer_nbytes(arr, check_min_size=None, cuda_support=True)
+            get_buffer_nbytes(arr, cuda_support=True)