--- a/ucp/_libs/utils.pxd
+++ b/ucp/_libs/utils.pxd
@@ -8,4 +8,6 @@ from libc.stdint cimport uintptr_t
 
 
 cpdef uintptr_t get_buffer_data(buffer, bint check_writable=*) except *
-cpdef Py_ssize_t get_buffer_nbytes(buffer, check_min_size, bint cuda_support) except *
+cpdef Py_ssize_t get_buffer_nbytes(buffer,
+                                   Py_ssize_t min_size=*,
+                                   bint cuda_support=*) except *