--- a/ucp/endpoint_reuse.py
+++ b/ucp/endpoint_reuse.py
@@ -98,10 +98,10 @@ class EndpointReuse:
 
         return core.create_listener(_handle, port=port)
 
-    async def send(self, buffer, nbytes=None):
+    async def send(self, buffer, nbytes=-1):
         await self.handle.ep.send(buffer, nbytes=nbytes, tag=self.tag)
 
-    async def recv(self, buffer, nbytes=None):
+    async def recv(self, buffer, nbytes=-1):
         await self.handle.ep.recv(buffer, nbytes=nbytes, tag=self.tag)
 
     async def close(self):