Makes a few adjustments to `get_buffer_nbytes` signature. Namely replaces `check_min_size` with `min_size`, which is now typed as `Py_ssize_t`. If `min_size` is not positive definite, this is the same as skipping the checking. Also provides some reasonable default values to `min_size` and `cuda_support`.