@jakirkham 2020-09-09
FWIW with this makes a small, but notable improvement in how long this check takes to run. Included a benchmark below just for the NumPy case. Though the CuPy case will show a similar improvement (as the check behaves the same independent of being on host or device).

Before:

```python
In [1]: import numpy
   ...: 
   ...: from ucp._libs.utils import get_buffer_nbytes

In [2]: s = (10, 11, 12, 13)
   ...: a_c = numpy.arange(numpy.prod(s)).reshape(s)

In [3]: %timeit get_buffer_nbytes(a_c, check_min_size=a_c.nbytes, cuda_support=True)
709 ns ± 1.64 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

After:

```python
In [1]: import numpy
   ...: 
   ...: from ucp._libs.utils import get_buffer_nbytes

In [2]: s = (10, 11, 12, 13)
   ...: a_c = numpy.arange(numpy.prod(s)).reshape(s)

In [3]: %timeit get_buffer_nbytes(a_c, min_size=a_c.nbytes, cuda_support=True)
679 ns ± 7.38 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```